PROJECT/docker/requirements.txt
flask
requests
pandas
numpy
pyspark
apache-airflow
boto3
pyathena
pyarrow
matplotlib
tableauserverclient
tableauhyperapi